omnigate_worker/requirements.txt
nodriver>=0.47.0
redis>=5.0.0
asyncpg>=0.29.0
SQLAlchemy[asyncio]>=2.0.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.1
tenacity>=8.2.3
httpx>=0.27.0
pyotp==2.9.0
pycryptodome>=3.20.0
